cleanup: cargo clippy --fix pass + manual triage of remaining lints
Sweeps through the lint backlog across all five feature configs
(`embassy`, `embassy-watch`, `embassy-game`, `embassy-mesh`,
`simulator`).  No behavioural change beyond what `clippy --fix`
introduces (which is itself behaviour-preserving by design); the
firmware passes a manual flash-and-poke (Maze, Tic-tac-toe, sprite
animation, factory reset, watch face).  +880 B flash on the full
build (`is_some_and` etc. inline slightly differently than the
`map_or(false, …)` pattern they replaced); RAM unchanged.

## Autofix (`cargo clippy --fix`, ~85 lints)

Mechanical, applied to both the `embassy` and `simulator` configs:

  * `collapsible_if` ×29 — `if a { if b { … } }` → `if a && b { … }`
  * `manual_range_contains` ×7 — `x >= 0 && x <= 23` →
    `(0..=23).contains(&x)`
  * `unnecessary_map_or` ×5 — `opt.map_or(false, |x| x > 0)` →
    `opt.is_some_and(|x| x > 0)`
  * `needless_return` ×12 (5 + 7 in simulator)
  * `needless_range_loop` ×3 (the easy ones — see below for the
    leftover)
  * `unnecessary_cast` ×3
  * `manual_div_ceil` ×3
  * `let_and_return` ×3
  * `let_unit_value` ×3
  * `manual_is_multiple_of` ×2
  * `identity_op` ×2
  * `unnecessary_mut_passed` ×1
  * a long tail of single-occurrence trivia.

## Manual triage (~17 more lints)

  * **`absurd_extreme_comparisons` ×8** — `state.<stat> >= STAT_MAX`
    on a `u16` field where `STAT_MAX = u16::MAX`.  Replaced with
    `== STAT_MAX`.  Same value (u16 can't exceed 65535); reads
    truthfully.  Two sites in `game::engine::{mod,to_display}`.
  * **`declare_interior_mutable_const` ×3** — `game::maze` was
    using the older `static X: [Atomic*; N] = { const INIT: …; [INIT; N] };`
    workaround.  Modernised to `[const { Atomic*::new(…) }; N]`,
    which is the inline-const form the rest of the codebase already
    uses for the same pattern.
  * **`empty_line_after_doc_comments` ×3** — `fw::board.rs` had a
    `///` comment on the wrong scope (was meant for the macro it
    documents) — converted to `//!` module doc and merged the
    surrounding comment fragments.  `game::maze.rs` had an
    orphaned `/// Generate a new perfect maze…` from a refactor;
    moved to its rightful home above `fn generate()` and demoted a
    section-divider doc comment to plain `//`.
  * **`doc_overindented_list_items` ×3** — fixed the over-indented
    `//!` list in `watch::ics.rs`.
  * **`needless_range_loop` ×3** — `contacts::hash_index_fetch`
    rewritten to `out.iter_mut().take(count).enumerate()` (clearer:
    iterate the destination, parameterise into the source by `i`);
    `tictactoe` draw loop now uses `board.iter().enumerate()` so
    `match cell` replaces `match board[i]`; `maze` exit-validator
    now `for &ev in &exit_cells[..already_placed]`.
  * **`manual_clamp` ×1** — `(x).min(MAX).max(1)` →
    `(x).clamp(1, MAX)` in `contacts`.
  * **`drop_non_drop` ×1** — removed an explicit `drop(sensor)` in
    `fw::temperature` (the embassy `Temp` type doesn't implement
    `Drop`, so the call was a no-op).  Comment updated.
  * **`never_loop` ×1** — `fw::mesh::persister::factory_reset_loop`
    had `loop { wait().await; wipe_and_reset().await; }` but
    `wipe_and_reset` returns `!` (it reboots), so the surrounding
    loop only ever executed once.  Dropped the loop; behaviour
    unchanged.
  * **`while_let_loop` ×1** — `sprite_loader::init` rewritten as
    `while let Ok(Some((name, _file))) = dir.next().await { … }`.
  * **`new_without_default` ×1** — added `impl Default for
    ContactStore` delegating to `Self::new()`.
  * **`wrong_self_convention` ×1** — `PetRecord::to_bytes(&self,
    buf)` → `to_bytes(self, buf)`.  PetRecord is `Copy`, no
    callsite changes.
  * **`needless_borrows_for_generic_args` ×2** — false positives
    inside the `#[gatt_service]` macro expansion in
    `fw::mesh::ble`; item-level `#[allow]` doesn't propagate into
    proc-macro output, so the suppression is at module scope with
    a comment explaining why.

## Deferred — wants explicit design

Not touched in this pass; each is its own change:

  * **`too_many_arguments` ×6** — `init_ble` (21), `init_epd` (12),
    `run_meshcore_listener` (10), `SimpleLoRa::new` (10),
    `try_handle_txt_msg` (8), `update_health_for` (8).  Each wants a
    config-struct refactor.
  * **`result_large_err` ×1** — `Result<(), TxRequest>` where
    `TxRequest` reaches 259 B.  Fix is `Box<TxRequest>`, but
    we're `no_std` without a global allocator.
  * **`needless_range_loop` ×1** —
    `watch::alarm::check_and_fire_alarm` iterates `0..N_ALARMS` and
    indexes a dozen parallel `Atomic*` arrays via accessor
    functions.  `enumerate()` over a single array would leave the
    rest still indexed; net wash on readability.

+++ src/lib.rs
@@ -701,7 +701,7 @@ where
 
     // Interior: 16×8 (body minus 2px border). Fill from right to left.
     let interior_w = 16u32;
-    let fill_w = ((pct as u32).min(100) * interior_w / 100) as u32;
+    let fill_w = (pct as u32).min(100) * interior_w / 100;
     let fill_x = bx + 2 + (interior_w - fill_w) as i32;
     if fill_w > 0 {
         Rectangle::new(Point::new(fill_x, y + 2), Size::new(fill_w, 8))
